+++ src/app/PostsList.tsx
@@ -3,9 +3,9 @@
 import Link from "next/link";
 import { useState } from "react";
 import TabFilter from "./TabFilter";
+import DraftToggle from "./DraftToggle";
 import { pretendard } from "@/app/fonts";
 import type { BlogPost } from "@/utils/githubBlogPost";
-import { useRouter, useSearchParams } from "next/navigation";
 
 interface TabViewProps {
   posts: BlogPost[];
@@ -22,25 +22,8 @@ export default function PostsList({
   showDraftToggle = false,
   initialShowDrafts = false,
 }: TabViewProps) {
-  const router = useRouter();
-  const searchParams = useSearchParams();
   const [selectedTag, setSelectedTag] = useState<string | null>(initialTag);
   const [showDrafts, setShowDrafts] = useState(initialShowDrafts);
-
-  const handleDraftToggle = () => {
-    const newShowDrafts = !showDrafts;
-    setShowDrafts(newShowDrafts);
-    
-    // URL 파라미터 업데이트
-    const params = new URLSearchParams(searchParams.toString());
-    if (newShowDrafts) {
-      params.set('showDrafts', 'true');
-    } else {
-      params.delete('showDrafts');
-    }
-    
-    router.push(`?${params.toString()}`);
-  };
 
   const filteredPosts = selectedTag
     ? posts.filter((post) => post.frontmatter.tag.includes(selectedTag))
@@ -56,41 +39,10 @@ export default function PostsList({
           totalPosts={posts.length}
         />
         {showDraftToggle && (
-          <div className="mt-6 p-4 border border-[var(--color-border)] rounded-lg bg-[var(--color-bg)]">
-            <label className="flex items-center justify-between cursor-pointer">
-              <span className="text-sm font-medium text-[var(--color-text)]">
-                Draft 포스트 표시
-              </span>
-              <button
-                type="button"
-                role="switch"
-                aria-checked={showDrafts}
-                onClick={handleDraftToggle}
-                className={`
-                  relative inline-flex h-6 w-11 items-center rounded-full
-                  transition-colors focus:outline-none focus:ring-2 
-                  focus:ring-[var(--color-primary)] focus:ring-offset-2
-                  ${showDrafts 
-                    ? 'bg-[var(--color-primary)]' 
-                    : 'bg-[var(--color-text-secondary)]'
-                  }
-                `}
-              >
-                <span
-                  className={`
-                    inline-block h-4 w-4 transform rounded-full 
-                    bg-white transition-transform
-                    ${showDrafts ? 'translate-x-6' : 'translate-x-1'}
-                  `}
-                />
-              </button>
-            </label>
-            {showDrafts && (
-              <p className="mt-2 text-xs text-[var(--color-text-secondary)]">
-                ⚠️ 개발 환경 전용 기능입니다
-              </p>
-            )}
-          </div>
+          <DraftToggle 
+            showDrafts={showDrafts} 
+            onToggle={() => setShowDrafts(!showDrafts)} 
+          />
         )}
       </aside>
       <div className={pretendard.className}>
